+++ RELEASE_NOTES.md
@@ -4,6 +4,17 @@ Bug fixes:
  - Fixed #174: SVG export of subplots in Edge and FF
  - Legend clipping
 
+### 1.5.26 (February 20, 2019)
+
+New features:
+ - Subplots can have a common title
+ 
+### 1.5.25 (February 19, 2019)
+
+Bug fixes:
+ - Fixed #174: SVG export of subplots in Edge and FF
+ - Legend clipping
+ 
 ### 1.5.24 (February 19, 2019)
 
 New features: